VPP: Port adds a VPP binding
This patch introduces a new VppBinding field attached to a bridge
or router simulation Port. This field is read from the topology
and it indicates that a port has a VPP binding as opposed to a
normal datapath binding.

Currently the VPP binding field includes only an interface name,
but this will be extended to include additional functionality
either as fields or methods.

Ref: MI-1317

Change-Id: Ie74981f3c854f5c0099ec06a61393200d98001d7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>