Update patch set 4
Patch Set 4:

(1 comment)

> Line 194:

Just a mistake.
I looked at midonet-cluster log file when I wrote this doc part. The log contains the following line:
...
/midonet/zoom/0/models/NeutronPort/fd9acd91-e6b5-4f84-8bf0-ff07443e90c2
/midonet/zoom/0/models/Port/61aafd9f-2faa-409d-83b4-6230a225f032
/midonet/zoom/0/models/Route/d322567f-eade-47ff-b014-2922e60e414b
...

The first line is an updating NeutronPort therefore I put it in the doc.
Actually it's not related to the patch. I'll fix it.

Patch-set: 4
Label: Code-Review=0