Update patch set 8
Patch Set 8:

Oh, right. That makes sense. Though now that I think about it, if we're mocking out the storage engine, why is it necessary to be able to handle the ID list in any order? Shouldn't we be able to guarantee a deterministic order? It's not like the translator code does any random shuffles of ID lists.

Patch-set: 8