Update patch set 1
Patch Set 1:

Probably the reason is that the class is imported from the same package to which the outer object belongs. "KeyType" is ambiguous because it can refer either to the inner class, which is imported, or the outer object, which doesn't need to be imported because it's in the same package.

Patch-set: 1
Label: Code-Review=0