MDTS: increase logging on minions when testing
This patch moves the stderr of the agent and its minion subprocess to
the service logs (those being separated by test case) so we can track if
the processes failed while executing a test. This patch also adds the gc
logs of the minion process to the debug logs which are gathered after
all tests have been run.

Change-Id: Ieca966ab30e6e436c872a1640560df180fc29efc
Signed-off-by: Xavi León <xavi.leon@midokura.com>