Install VPP from packages on sandbox
This patch installs our vpp_lite binary package published from our
unstable repos in the midolman image. This version of VPP does not
use DPDK. The automatic start up of the vpp process by upstart is
disabled on this image. The VPP process should be managed by the
MidoNet Agent.

Change-Id: I454090e73169a4ce2157d1f6e4f2dc267cd3f104
Signed-off-by: Xavi León <xavi.leon@midokura.com>