Port mapper includes the port tunnel key
Currently the agent loads the port bindings in two steps: (i) first
it loads a Host device via the host mapper that loads the ports
previous host identifier and an interface name, and (ii) that Host
is converted into a ResolvedHost where the ports' tunnel keys
are asynchronously loaded via the virtual topology.

This second step is redundant.

This is a first patch towards elimiating this redundancy by loading
the ports' tunnel keys in the host mapper.

Change-Id: If2508221de13f4c4c4abda581c0d6958750fb9b6
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>