BGP: Routing handler sets the namespace name
Currently the bgp helper script takes many optional arguments
depending on whether the BGP namespace is setup for an interior or
an exterior port. This increases the complexity in the helper
script by having to handle several different cases, depending on
whether certain arguments are set or not, which in turn modify
the namespace or the interface name.

With this change the namespace and the interface name are set by
the routing handler, that uses either the current port interface
name if the port corresponds to a container (BGP for interior
ports) or it generates a name using the current port identifier
if the port is an exterior port.

Fixes: MI-1385

Change-Id: I53b77cd6a1e3ff0e19706c7a1b2313f3717f4e0b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 822605d6c3e44917d4a219020f3eaba6ee3bcd8e)