Utils: Introduce monitored processes
This patch factors out the handling of failures on subprocess so it can
be reused in later patches. It also fixes an issue with the restart
logic: if a process reaches the maximum number of restarts even if it is
in a long period of time, it will shutdown the Agent completely. This
patch improves the restart logic by only considering a time window for
failures to be accounted for. In other words, the monitored process
will only be restarted if it reaches the maximum the number of failures
within a specific time window.

It also fixes a small issue in the mocked unix clock which was defined
as a val, and thus reused by different tests. Depending on the order,
the assumptions that the clock starts at 0 was broken. This patch
allocates a new mocked clock for each test.

Change-Id: I4a4368e9a7900a22bb664c7b8a44a91cb12a3f28
Signed-off-by: Xavi León <xavi.leon@midokura.com>