Utils: Introduce monitored processes
This patch factors out the handling of failures on subprocess so it can
be reused in later patches. It also fixes an issue with the restart
logic: if a process reaches the maximum number of restarts even if it is
in a long period of time, it will shutdown the Agent completely. This
patch improves the restart logic by only considering a time window for
failures to be accounted for. In other words, the monitored process
will only be restarted if it reaches the maximum the number of failures
within a specific time window.

Change-Id: I35d5796f58e2a58dbb3f9678fcfe13ae5f2211fc
Signed-off-by: Xavi León <xavi.leon@midokura.com>