Update patch set 8
Patch Set 8:

Hi,
Is there is a command to restart "Midokura Bot(building and tests)", for example like "recheck" in OpenStack projects?
I've got two different unsuccessful results(Unstable, Failed) for the two latest patch sets but I cannot to reproduce it on my system. Is it possible to get logs? Or instructions how to reproduce it.

Patch-set: 8