Translators: Separate manager code and use transactions
Currently the C3POStorageManager has a dual purpose: it serves as
the manager for both the Neutron traslators as well as the Neutron
importer, which reads Neutron changes directly from SQL and applies
them to the NSDB.

The latter performs on initialization several importer specific
tasks, such as creating the last process task node in ZooKeeper,
tasks that are no needed for the operation of the Neutron
translators when used via the REST API.

This change separates the translator code from the importer code
in a new NeutronTranslatorManager class, that provides only the
initialization for the translator class mapping. This ensures that
importer specific code is not executed by components that don't
require it.

The patch also changes the interface of the manager's translate
method to use transactions when applying the result of a
translated Neutron operation to NSDB.

Ref: MI-1447

Change-Id: Ib42f179db3a925af19728c26dc907f5c1ab8b4e0
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>