Translators: Use transactions in translator interface
This adds a transaction argument to the translate methods from
the Translator interface, and updates all implementing classes.
This change ensures that a translator can use the transaction to
read ZOOM objects that will be modified during the transaction,
thus guaranteeing those objects either remain unchanged until
the transaction is committed, or if the objects change the
transaction will fail.

Ref: MI-1447

Change-Id: I8a9c740a6f8e01e3c2fbff3bf5197e270600dca1
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>