ZOOM transaction adds exists and unit tests
This adds a cacheable exists() method to the ZOOM transaction.
The value returned by this operation is consistent with any
operations previously performed as part of the same transaction,
i.e. the method returns false if the object was deleted during
the transaction but the transaction was not yet committed.

Calling exists() will also update the transaction cache such that
future calls to the exists() or get() method for the same object
does not incur an additional trip to ZooKeeper.

The method will also fail if the object is modified since the
creation of the transaction object.

Change-Id: I0e81369879a74efb0d9f6c339250b679fd0b4981
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>