Disable l2insertion test for PoC
The l2insertion test uses a fake snort that just bounces back everything
it gets. If service ports can receive non-redirected traffic, the fake
snort needs to be smarter (it needs to not bounce ARP requests for
example). Since this is just for a PoC, there's no point of changing
fake snort in all images, so I'm disabling the test in this branch.

I've manually tested with a modified fake snort that only bounces VLAN
tagged packets, and the test passes with this.

Change-Id: I6b55c8ee53b60141ec031a5d92e4d2850fa6b48f
Signed-off-by: Ivan Kelly <ivan@midokura.com>