VPP: Midolman kills stale vpp processes
This patch modifies the vpp-start script that midolman uses
to launch vpp so that already running vpp instances are killed
before vpp is launched.

This is needed because otherwise vpp will fail to run due to
the listen port being in use, and failure to launch vpp will
result in midolman termination.

Change-Id: Iddafa727cc118ba0eacd043e46fc9536b4508e6c