MDTS: install dependencies on top of sandbox
We were installing MDTS dependencies before installing sandbox to
override some sandbox dependency versions. This doesn't make sense as
MDTS works on top of sandbox, so it should work out of the box
installing in the same virtualenv. This patch removes duplicate
dependencies in MDTS so we rely on the sandbox counterparts.
Also, configure sandbox to pull from artifactory v2 registry, which is
now possible with newer versions of docker.

Change-Id: Iecbb338cdb8639f634be937ccfba7361ef83aacc
Signed-off-by: Xavi León <xavi.leon@midokura.com>