FIP6: Add back-channel flow messages to flow processor
This moves the back-channel flow messages from the packet workflow
to the flow processor, where these messages originate. These will
be used, in addition to the packet workflow, by the VPP controller
to install VPP flows.

In addition, the flow processor sends a message upon flow error
that notifies the VPP controller that it failed to install a flow.

Ref: MI-1528

Change-Id: I7ceedfbd06894b0abe63c77f777f6d6725158c80
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>