FIP6: Start DP channel with DP service
Previously, the datapath channel was started by the packet workers
service. In order to allow the VPP controller to leverage the
same datapath connection, the channel with the underlying flow
processor must be started before, since the VPP controller may
send events before the packet workers are initialized.

This change modifies the initialization of the datapath channel
from the packet workers service to a new datapath service, which
has been renamed from the datapath connection service. This
ensures that the channel disruptor is stared before the VPP
controller.

Ref: MI-1528

Change-Id: Ieb0ed7daa3cef60465fb39e5930d79d5165e9b23
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>