FIP6: Add a callback-based back-channel shard
Currently, the simulation back-channel uses queue-based back
channel shards, that allow the packet workers to poll back-channel
messages. However, such a back-channel notification mechanism may
be too expensive for the implementation of the VPP controller,
which only needs to process flow operation result back-channel
messages. These are expected to be infrequent and specific.

This patch adds a new back-channel shard that uses a callback
function to send back-channel messages to consumers. This will
allow the VPP controller to hook its own callback function to
the back-channel, filter unwanted messages, and only process
those that are specific to the VPP controller.

For a callback back-channel shard, the callback function should
perform a non-expensive operation, such as filtering the back
channel message, and then enqueue them onto an executor or send
them to an actor.

Ref: MI-1528

Change-Id: I347fdb3d7a74790cb4d294ccbedba69aad452772
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>