API: Cleanup storage usage
This makes the storage private in the MidoNet resource class, and
removes the usage of storage from several resources that were
still using it. This ensures that all calls to storage are made
either through the resource transaction or by calling methods in
the base resource class.

The patch also removes other unused code.

Change-Id: I6d40402fba4a6d1240ef62301a35827d1363438e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>