Update QoS API to allow rules with policy
Update the models to add a "rules" section so that requests can
include the whole rule list with the policy.  This list will also
be stored as informational reference with the policy object (it
will not be used otherwise, after the model is stored).

If the rules are sent with the policy, run through the list and
with add rules (the rule(s) or policy was not present previously),
delete rules (rules were present previously but are not in the
policy object being posted), or update rules (the rule existed
previously and exists in the current policy model, but the data
in the rule has been modified).

This will update the midonet model rule objects stored at the
top level and add or remove references from the policy's ID
list as well, when applicable.

If the rules element is not sent into the API, the rules
will be left alone.  If it is sent in and is an empty list,
then the rules will be deleted on the policy.

After the rules are used to create the top level objects, they
are discarded and any response will have a null "rules" field.

Added apprproriate unit tests for these cases.

Change-Id: I391f54598b7842556fa01812dfa84e15500d9aa7