MDTS: specify sandbox name in args
This patch allows users to specify the name of the sandbox
in which the tests should be run on. This can be done in two
new ways:
1) By passing an additional `-n <name>` option to the run_tests
   script.
2) By exporting an env variable `export MDTS_SANDBOX_NAME=<name>`

Previously this needed to be changed in the mdts.conf file,
now this optional argument superseeds this last option.

Change-Id: I725be588b7fea4c4e52a2212153b730b6f2bf9f1
Signed-off-by: Mateo Hermida <mateo@midokura.com>
(cherry picked from commit d43605fc4953a383a25eb6565d1482382cb045e5)