Fix flake state proxy test
The flake test referenced in this patch checked for the existence of a
state key in zookeeper before actually happening. Because the delete
operation is asnyc after an add request, we do not guarantee that the
key is removed after the local cache receives the read callback.

This patch just waits on the non existense of the removed state key that
eventually should be removed.

Ref.: MI-1574
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Ibb210c8c0711671977e87959a70a180919b78218