Show L4 range start-end even if they are the same
In midonet-cli, a L4 port range is shown in the <start>-<end> format
unless the <start> <end> values are the same, in which the value is
present.

But since this values are also used to show the values for ICMP type
and ICMP code, they always have different meaning, despite it can be
the same number, so collapsing it masks one of them.

This commit removes the collapsing, so both values are shown always.

Ref: MI-642

Change-Id: I76847c5f2781131c445424b78ca32109d847e4fe
Signed-off-by: Miguel Herranz <miguel@midokura.com>
(cherry picked from commit eb658deddae786174ab82c48c00e43997a801b2d)