FIP6: Update interace names and remove some warnings
This changes the interface names used for hooking the uplinks via
VPP to the format vpp- or ovs- followed by the first 8 hex digits
of the port identifier. This is to be consistent with the inteface
names we used for BGP or VPN, where the interface name prefix is
bgp- and vpn-, respectively.

The patch also removes the call of .size on arrays to .length,
since the former requires an implicit conversion of the array to
a scala sequence.

Ref: MI-1589

Change-Id: Ie8dfe8d016c0ce4cf65c5377a180db2f2e15c862
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>