Add QoS Service
This is a new host service that runs on all midolman hosts
and translates QoS configurations to linux TC netlink requests.

It works by listening to InterfaceScanner for the information
on currently UP interfaces on the local host. It combines this
information with the topology information about the ports that
are currently bound to interfaces on the local host. If a match
is found (i.e. there is a port bound to a local interface that is
UP and has QoS policy information configured) it will pass on the
request to a thread that passes messages on to the kernel.

Signed-off-by: Joe Mills <joe@midokura.com>
Change-Id: Id4acbd1b45cf7f5fc9b5180f5a0635555c8fec0e