Make ZOOM transaction delete node idempotent
The createNode and deleteNode methods were introduced in ZOOM to
allow the Neutron translators to make low-level modifications in
ZooKeeper for paths that are not part of the ZOOM topology, mainly
state table entries.

However, the translators operate under the assumptions that
deletions are idempotent (i.e. they do not fail if the path to
be deleted does not exist). This is not so for the deleteNode
method.

This patch introduces a new `idempotent` argument to the deleteNode
method, which defaults to true, allowing the caller to select
whether the deletion is idempotent.

Fixes: MI-1600

Change-Id: Iced6eec8f0775a23a72ec22b627747d603434b83
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>