Update patch set 2
Patch Set 2: Code-Review-1

It's unclear how this will play for BC. If you have an old install (HM or whatever) expecting the lock, and the new code doesn't use it, then you'll end up with conflicts (the new code will be able to modify nodes, while an old code modification is in flight). Is it possible to check if the lock node exists, use the lock if it does, and skip it if it does not (i.e. does the lock node exist when the lock is not held)?

Patch-set: 2
Label: Code-Review=-1