Add support for transaction lock and retry to storage
This consolidates the transaction lock and retry used in several
parts of the code in a new method added to the Storage trait.
The new method will apply the lock and retry policy specified
in the backend configuration.

If a lock for the ZOOM topology exists or is created in ZooKeeper
all transaction operations require acquiring a lock before they
are executed. Otherwise, the transactions are lock-free, and
contention is governed by the retry policy.

Ref: MI-1503

Change-Id: Iee69d466517271e6ea58d39d0b9881b5516b9314
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>