Check for existing setup on port available
I observed some behaviour, where a port was bound twice and the vpp
controller would jam because it would try to set up the port again and
fail when it tried to add the ipv6 flows, and the teardown would break
the first setup and the situation would never be resolved.

This patch changes the initialization sequence to check if there is
already a port with that id set up, and if so, tear it down before
trying to set up again.

Change-Id: Ie043aeef249efaad5923d883a1914e7bffbbfb93
Signed-off-by: Ivan Kelly <ivan@midokura.com>