Added QoS MDTS lib and tests
Added QOS policy to VTM and as a model object for MDTS to use.
Added various functions to parse the QOS policy info from a
topology YAML file as well as add the obejcts programatically.

Joe Mills <joe@midokura.com> supplied the original code base for
the MDTS lib changes.

Verify bandwidth function is just a stub for now, to be filled
out in a subsequent patch.

Make sure these tests don't run in gates to start with.

For QOS policy tests, used test plan:

https://docs.google.com/a/midokura.com/document/d/1f9_I2tdKTN9XWzUh4qKjlolr4vww7gJAAU3g1zH4HXg/edit?usp=sharing

Change-Id: I421dc9952c867acb8ba732121da4709960c9eae4