Make CI use prebuild sandbox images and venvs
The CI images has been modified to run once the latest sandbox and
midonet, so the pip/gradle/docker images are precached in the image.

The usage of pip/gradle/docker cache should be transparent, but in
the case of docker, the pull of the images should be no longer needed.

The virtual envs are now located at /opt/midonet-venv and
/opt/midonet-sandbox-venv.

The version of the sandbox in the CI image is checked against latest,
so in case it is old, a new build of the images is triggered.

Change-Id: I61f79c69a2f1ffbc7fe3b89c5f2f7d400327ed3e
Signed-off-by: Miguel Herranz <miguel@midokura.com>