FIP64: Introduce VppDownlink
This adds a new VppDownlink trait for the VppController actor,
which is responsible for managing the current downlink ports and
their associated FIPs.

Upon start, this trait monitors the FIP64 state table in order
to discover the downlink ports associated with IPv6 floating IPs.
For each new port, the trait will create a DownloadState that
maintains the state for that port by monitoring the port via
the virtual topology and its associated FIP64 rules.

For each loaded port, the trait will send a set of messages
to self, as follows:

* CreateDownlink - indicates the addition of a new downlink port,
  where the message will include the port identifier, associated
  VRF table, port IPv4 and IPv6 subnets.

* UpdateDownlink - indicates that the configuration of the
  downlink port has changed. Only the port's IPv6 address can
  change.

* DeleteDownlink - indicates the deletion of the downlink port,
  where the message includes the port identifier and VRF table.

* AssociateFip - indicates the addition of a FIP64 for a given
  downlink port, where the message includes the port identifier,
  VRF table, floating IPv6 and fixed IPv4.

* DisassociateFip - indicates the deletion of a FIP64 for a given
  downlink port, where the message includes the port identifier,
  VRF table, floating IPv6 and fixed IPv4.

Ref: MI-1665
Ref: MI-1659

Change-Id: Iaed656af907795af0d354a4e98513216b9affbfc
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>