FIP64: Serialize VppController operations
This ensures that all opearations are serialized on the conveyor
belt. To this end, all operation methos return a future, such that
progress to a subsequent opeation notified to the actor cannot be
made until the previous operation completes. This ensures that
a new opeation, that is due to start due to a change in the
topology, does not find the intermediate state generated by
another in-progress operation.

The patch also re-organizes the controller methods, such that they
are now divided between: attachUplink, attachDownlink, detachUplink
and detachDownlink.

Ref: MI-1711

Change-Id: Ic39511036a8c53c5276e99e9459aacc460eb8696
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>