FIP64: VPP controller handles uplinks and downlinks
This allows the VPP controller to manage uplinks and downlinks
separately. This is because the VPP controller should stop
monitoring the downlinks if there are no uplinks.

To this end, the patch introduces a LinksMap type, which is a
Java hash map, allowing the corresponding uplinks to downlinks
map to be passed as argument to the `attachLink` or `detachLink`
method.

The patch replaces the Scala map with the Java map for being more
lightweight and having method such as remove(key, value).

Change-Id: I1034abf91b7785ecf60650360c495fc68ff10cf7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>