Router Interface Translator supports IPv6
This allows the router interface translator to handle the
creation of gateway ports with IPv6 addresses. Upon the creation
of such a port the translator will:

* Assign the router port address/subnet to an available link-local
  IPv4 subnet (169.254.x.x/30).

* Assign a tunnel key to the port, because the port will be bound
  to a physical interface.

* Modify the port route to match the target NAT64 IPv4 pool
  instead of the port's subnet.

* Create a NAT64 rule with a deterministic identifier that contains
  the port's IPv6 address/subnet and the IPv4 NAT64 pool.

The router port created by the router interface translator is
deleted from the port translator when its corresponding peer port
is deleted. This translator doesn't need to be changed because
the only addition, the NAT64 rule, is removed by cascade deletion
due to the ZOOM bindings, when the router port is deleted.

Ref: MI-1670
Ref: MI-1622

Change-Id: I387bd14132048d9bb3a2f7ca8efd7962634d74fb
Signed-off-by: Adrian Serrano <adrian@midokura.com>
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>