Set TOS field of packet if DSCP is set
Change the TOS field of any IPv4 packet if the DSCP rule is set
on the port.  This should affect the simulated packet as well as
be added as an action on the datapath for subsequent packets.

Mark the TOS field as "seen" if the DSCP is found, which will
ensure that the TOS field will be used to match the flow, so
other packets with a different TOS field will miss the match
and resimulate (this is to prevent a packet with a TOS field
set to the same value as the dscpMark from installing an
incorrect flow which will match ALL TOS values, as this
particular flow will not have a setKey action (because the diff
won't show a changed value, and hence won't install an action
to affect the value), so instead targeting this flow on that
particular TOS value will ensure that packets with other TOS
values will get resimulated and a setKey action WILL be added
in those cases).

Only set DSCP on IPv4 packets (not ARP, etc.)

Added unit test for DSCP marking.

Change-Id: Ie255b972bfc3fe60cd6cd7782a830ad36f68177e