Set TOS field of packet if DSCP is set
Change the TOS field of any IPv4 packet if the DSCP rule is set
on the port.  This should affect the simulated packet as well as
be added as an action on the datapath for subsequent packets.

Add a flag to FlowMatch (of which type PacketContext.wcmatch is)
to show that the tos key was reset, so that even if the incoming
packet's ToS field is the same as the DSCP mark, the setKey
action will still get added to the flow.  Just using a diff
wouldn't work in this case as the key wouldn't be added.

Only set DSCP on IPv4 packets (not ARP, etc.)

Added unit test for DSCP marking.

Change-Id: Ie255b972bfc3fe60cd6cd7782a830ad36f68177e