Set TOS field of packet if DSCP is set
Change the TOS field of any IPv4 packet if the DSCP rule is set
on the port.  This should affect the simulated packet as well as
be added as an action on the datapath for subsequent packets.

Check during port ingress if a) QOS policy is set and b) if the
policy has DSCP rules, in which case alter the TOS field of the
wcmatch with the mark.  Also add a check in the diffIP function
to also check for difference in the TOS field.

Change-Id: Ie255b972bfc3fe60cd6cd7782a830ad36f68177e