Add MDTS test for IPv6 FIP create using neutron API
This commit add test_neutron_fip() test to test the changes in the
Midonet Neutron plugin that allow to create a IPv6 FIP.

NOTE: the sandbox has to be updated so it have the latest changes that
make sandbox to override upstream Midonet Neutron plugin with the one
in midokura/networking-midonet repo. Also, other changes from cluster
are needed for the test to pass. Mark test as nottest to not break
nightly regressions

Ref: MI-1660

Change-Id: Ibd9fe0cf99997543f262a0a975c483f96ce25bd5
Signed-off-by: Miguel Herranz <miguel@midokura.com>