Check if entry existed in ArpCache before using it
If two arp cache entries are added for an IP with two different MAC
addresses, then two expiries will be queued, while there is only ever
one arp cache entry. This can lead to a case were the second expiry
tries to expire an entry that doesn't exist. This patch fixes this, by
checking if the entry existed before using it.

Ref MI-1716

Change-Id: Ic6d453ff957b39fad6d99b49f8e00312cb594dca
Signed-off-by: Ivan Kelly <ivan@midokura.com>