Retry Keystone authentication on unauthorized
Currently, the MidoNet Cluster only authenticates during the
first client request, and subsequently only when the cluster
token is about to expire. This causes the cluster to fail
performing Keystone operations if the token has been invalidated,
without having expired.

This enables the cluster to retry authenticate if an operation
fails because the cluster authorizarion fails. The cluster will
only retry once per operation.

Ref: MI-1654

Change-Id: Id0c30b8a424a989970c7420a5e4032a54acdadc4
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>