QOS bandwidth rule burst defaults to 80% of rate
If the user does not specify the burst parameter when creating a QOS bandwidth
limit rule, the QosPolicyMapper will set the burst to 80% of the maxKbps
parameter, i.e. to 0.8 seconds' worth of bandwidth allocation. By design, this
is not reflected in Zookeeper or visible to the user via API queries.

Change-Id: I8737c3508951186ed03ff829d2eda4a8f2ea6132
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-1687