FIP64: Router translator manages IPv6 gateway ports
This modifies the RouterTranslator to allow CRUD operations for
IPv6 gateway ports. To this end the patch verifies the gatewway
port IP address and adds custom methods to manage IPv6 gateway
ports.

For every IPv6 gateway port, these will configure routes to the
VPP IPv4 address, and the NAT64 FIP rule, such that the VPP
downlink is able to process these ports.

TODO: Some of the tests are commented-out because they require
additional changes to the SubnetTranslator.

Change-Id: I65c36a986b17021b4da86e124bf63256abc71649
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>