FIP64: Make router gw port deletion idempotent
The deletion fails because the translator performs a get on the
port, even if the deletion operation itself is idempotent.

This occurs when the gateway is first unset on the router, which
will delete the corresponding router port, and then the Neutron
port is deleted.

TODO: The PortTranslator should support IPv6 to cleanup the
corresponding IPv6 objects, if the port is deleted first.

Change-Id: I8a41c63c801cdbce513f6414f2d45fd9489b7eef
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>