FIP64: Floating IPv6 translator uses correct port
This fixes an issue in the Floating IPv6 translator, where the
translator was adding an entry to the FIP64 using the identifier
of the port to which the FIP6 was associated (usually an instance
port) rather than the gateway port on the tenant router.

The VPP controller requires the latter in order to determine to
which downlink port the FIP6 applies.

Change-Id: Iccb5231c2b2a42a0f926a4cabfcaccccaf68320d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>