Enable topology lock by default
This ensures that the MidoNet backend enables the topology lock
by default. The lock is enabled from that backend rather than
setting the lockFree flag to false to maintain the test coverage
for the existing code.

The lock is re-enabled because the API operations have a high
probability of failures during calls with a high degree of
concurrency. It may be disabled again in a future patch, if
this issue is resolved.

Ref: MI-1739

Change-Id: Icc29ab83044210ed2ab033f10983d1f03dbac5cb
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>