FIP64: Fix downlink bound to empty interface
This fixes an issue where a downlink was bound to an empty
interface, and the datapath controller remained stuck on this
empty interface when updated.

The cause for this was that the VPP controller did not clear
either the interface name or the host identifier, when undoing the
binding.

With this change the VPP controller will set the binding fields
when the binding is created, and clear them during cleanup. It is
not needed to store the previous interface name, since the VPP
controller alone manages the bindings for the downling ports and
has the authority to modify the bindings overwriting previous
configuration. Restoring previous configuration is not needed.

Change-Id: I9cda209b9e99d710f77fefa9afe7db47612dde1d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>