Keystone: Fix static admin token cleared on unauthorized
This fixes an issue in the Keystone client, where the admin token
using the admin token configuration (rather than admin username
and password) was being set to null during a Keystone unauthorized
exception. This caused the cluster to fail to authenticate itself
with Keystone for subsequent calls if the admin credentials were
missing in the configuration.

The patch ensures that the cluster retries to authenticate itself
only for tokens obtained with username/password credentials. Calls
that used the static admin token do not have to be repeated, and
the token doesn't have to change.

Change-Id: I756016398da9b869e51ed1f4cbf11eabe6df3859
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>