Prepare router translator to support IPv6
This groups code specific to IPv4 from the router translator to
specific methods, such that the translator can be adapted to
enabling IPv6. To this end, the patches introduces IPv4-specific
methods for create, update and delete.

The patch enahances the handling of the router gateway port, and
in particular the handling of the router updates, by comparing
the old and new Neutron routers and determinining what has been
updated.

The changes also include removing certain redundant operations,
such as reading of the router peer port, which was used to
determine during a router update whether a new gateway port was
added.

Change-Id: I4bc547a608b862b0066dce1d7adc3e7f5735ea4d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>