Update patch set 4
Patch Set 4:

Ernest, do you know which fields are null or what packet are you processing when this happens? It seems to me that the fields that don't match a flow key when a packet it received, then that field is left to null. For example, a L2 packet won't have the L3 fields set.

Patch-set: 4