FIP64: Verify router has gateway port during update
This fixes an error in the router translator update method, where
updating a router with no previous or current gateway port, would
check whether the port `snat_enabled` field has changed. This
should only be the case if the router has a gateway port, and this
additional condition was missing.

Change-Id: Ic9afa5c342fa1043afcd288263c554d3a4f9985a
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>