FIP64: Add feature flag for VXLAN downlink
This adds a configuration flag that indicates whether FIP64
downlinks use VXLAN. The configuraton is passed as a system
property, since this is an internal configuration (should not
be available in the mn-conf schema).

The configuration is also static, meaning that we don't have to
add an instance of MidolmanConfig to every simulation class where
this might be needed.

Change-Id: I2deab32ba750afc7c6fb771ab3c9cf2241d16485
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>