Wait vpp to finish initialization to avoid races
Vpp process in lunched by midolman. Tests
need to make sure the VPP is running and the
initialization is done before adding default
routes to vpp.

This patch also modifies private net range for
neutron test. It used to be equal to the pool range
of VPP causing midolman drop the packet with src IP addr
been equal to the subnet gateway.

Change-Id: If53fa8924bbd8eb66abda0684f90e620a2320056
Signed-off-by: Dmitry Bufistov <dmitry@midokura.com>