Fix container scheduler fails after retries
This fixes an issue in the container scheduler where the
triggering of an internal timer, such as a retry timer or
scheduling timer, may cause certain subsequent operations to
fail because the observable of the feedback loop is in completed
state.

The cause of the issue is that the scheduler did not insulate
the feedback loop from the completed notification of `timer`
observable, causing it to become inoperable when the first timer
would trigger, such as a scheduling timeout or a scheduling
retry. Although the scheduler would still respond to other
events, such as changes in the topology, all events that would
have used the feedback loop were subsequently blocked.

This fix adds an intermediate feedback observer that forwards
onNext and onError notifications, but blocks onCompleted
notifications that are expected for timer observables.

Ref: MI-1736

Change-Id: I8e543f4197033e3e92c64a2af0c87f9a6cadef98
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>