Add backpressure to container service observables
This increases the capacity of the backpressure buffers used
by the container service from 0x1000 to 0x100000, which represents
the maximum number of notifications that will be cached before
the service observable terminates with onError.

The patch also introduces backpressure to several observables
that use `observeOn`, to ensure that notifications do not
overflow the `observeOn` buffer, which defaults to 128.

The patch also removes several redundant `observerOn` operations
where the original notifications are known to be already
scheduled on the context scheduler.

Note: The capacity of the backpressure operator represents only
a limit to how many notifications are cached. The underlying
concurrent queue ensure that only the queued notifications account
for memory usage.

Change-Id: I3f968a24f44bc0148f1d8847fdaa01e0a3da0e4c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit 527989b1f84d6dace3309d95a5afa6dc05acc760)