Change authentication error log message to info
When a client fails to authentication, normally when using an
expired token, the REST API logs an error with a little
information and a complete stack trace. However, these types
of authentication errors are common.

For this reason, this patch downgrades the severity of the log
message to info, removes the stack trace, and includes the
remote address of the client that failed to authenticate.

Change-Id: I3f3e2b854c4700a9d02d83826a2ec2411de96cf9
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>