Prepare IPSubnet to handle address and subnet
This changes the implementations of the IPSubnet interface to
allow them to store both an IP address and subnet prefix length.

In addition, the patch includes the following:

* Removes unused features such as JSON annotations that were only
used for v1 ZooKeeper serialization.

* Adds toNetworkAddress() and toBroadcastAddress() methods to
the interface and provides implementations for IPv6.

* Enhances unit tests for IPv4 and introduces tests for IPv6,
providing 100% coverage for these classes.

* Fixes a bug where IPv4Subnet.toBroadcastAddress() would return
an incorrect result if the subnet prefix was 32.

* Modifies the string parsing methods to use more lightweight
Guava methods rather than regular expression matching.

Change-Id: I6a5892104b49dbe09791e563945b5c437aa41dfc
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>